User data storage created.
App uses window.localStorage to save
-name
-values
-profile picture
-todos
-habits
-level
-exp bar

If the userdata does not exist (as in first time using the site) a modal
will pop up asking for a person's name and values.

Values are limited to 6. The app will warn someone when they've gone
over and tell them which one was deleted.

Saving between sessions now exists.